100-Days-of-Code-The-Complete-Python-Pro-Bootcamp-for-2023/Section-64-Day-64-Advanced-My-Top-10-Movies-Website/day-64-starting-files-top-movies/requirements.txt
Bootstrap_Flask==2.2.0
Requests==2.31.0
WTForms==3.0.1
Flask_WTF==1.2.1
Werkzeug==3.0.0
Flask==2.3.2
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.25

python-dotenv==1.0.1